feat : `Splash`에서 처리하던 화면이동처리를 `Root`로 이동시키고 `Splash`, `SignIn`, `Home`을 `Root` 하위 수평관계로 변경했어요
