+++ SixthSense/VegannerApp/Sources/Root/RootRIB/RootRouter.swift
@@ -10,8 +10,12 @@ import RIBs
 import Account
 import Then
 import Splash
+import Home
 
-protocol RootInteractable: Interactable, SplashListener {
+protocol RootInteractable: Interactable,
+                            SplashListener,
+                            SignInListener,
+                            HomeListener {
     var router: RootRouting? { get set }
     var listener: RootListener? { get set }
 }
@@ -21,15 +25,21 @@ protocol RootViewControllable: ViewControllable {}
 final class RootRouter: LaunchRouter<RootInteractable, RootViewControllable>, RootRouting {
 
     private let splashBuilder: SplashBuildable
+    private let signInBuilder: SignInBuildable
+    private let homeBuilder: HomeBuildable
 
-    private var childRouting: SplashRouting?
+    private var childRouting: ViewableRouting?
     
     init(
         interactor: RootInteractable,
         viewController: RootViewControllable,
-        splashBuilder: SplashBuildable
+        splashBuilder: SplashBuildable,
+        signInBuilder: SignInBuildable,
+        homeBuilder: HomeBuildable
     ) {
         self.splashBuilder = splashBuilder
+        self.signInBuilder = signInBuilder
+        self.homeBuilder = homeBuilder
         super.init(interactor: interactor, viewController: viewController)
         interactor.router = self
     }
@@ -41,7 +51,7 @@ final class RootRouter: LaunchRouter<RootInteractable, RootViewControllable>, Ro
     }
 
     
-    private func attachSplash() {
+    func attachSplash() {
         if childRouting != nil { return }
         
         let router = splashBuilder.build(withListener: interactor)
@@ -51,6 +61,41 @@ final class RootRouter: LaunchRouter<RootInteractable, RootViewControllable>, Ro
         viewControllable.present(viewController, animated: false)
         attachChild(router)
         self.childRouting = router
-
+    }
+    
+    func detachSplash() {
+        guard let router = childRouting else { return }
+        router.viewControllable.dismiss(animated: false)
+        detachChild(router)
+        self.childRouting = nil
+    }
+    
+    func attachSignIn() {
+        if childRouting != nil { return }
+        
+        let router = signInBuilder.build(withListener: interactor)
+        let viewController = router.viewControllable
+        viewController.uiviewController.modalPresentationStyle = .fullScreen
+        viewControllable.present(viewController, animated: false)
+        attachChild(router)
+        self.childRouting = router
+    }
+    
+    func detachSignIn() {
+        guard let router = childRouting else { return }
+        router.viewControllable.dismiss(animated: false)
+        detachChild(router)
+        self.childRouting = nil
+    }
+    
+    func attachHome() {
+        if childRouting != nil { return }
+        
+        let router = homeBuilder.build(withListener: interactor)
+        let viewController = router.viewControllable
+        viewController.uiviewController.modalPresentationStyle = .fullScreen
+        viewControllable.present(viewController, animated: false)
+        attachChild(router)
+        self.childRouting = router
     }
 }
